core: Switching to 32bit vertex indices by default ...
TODO: Test on mobile and web

+++ core/src/chr/gl/BatchMap.h
@@ -39,7 +39,7 @@ namespace chr
       std::map<Texture, VertexBatch<V>> map;
     };
 
-    template<int V = XYZ, typename I = GLushort>
+    template<int V = XYZ, typename I = GLuint>
     class IndexedBatchMap
     {
     public:
